Add LLM::Provider#models
This new method returns a Hash where the key is the model name,
and the value is an instance of LLM::Model. Each LLM::Model describes
the character of a model, and it can be used in place of a model name.
For example:

llm = LLM.ollama(nil)
convo = llm.chat "How are you?", model: llm.models["deepseek-r1"]

The other notable change included in this diff is that when you start
a conversation with a given model, that choice is remembered throughout
the conversation (along with any other parameters that the conversation
is created with).